+++ s3torchconnectorclient/rust/src/exception.rs
@@ -5,6 +5,7 @@
 
 use std::error::Error;
 use std::fmt::Write;
+use log::error;
 
 use pyo3::exceptions::PyException;
 use pyo3::PyErr;
@@ -14,6 +15,10 @@ pyo3::create_exception!(
     S3Exception,
     PyException
 );
+
+fn log_error(message: &str) {
+    error!("ERROR: {}", message);
+}
 
 pub fn python_exception(error: impl Error) -> PyErr {
     let mut s = String::new();
@@ -25,7 +30,10 @@ pub fn python_exception(error: impl Error) -> PyErr {
         write!(&mut s, ": {}", error).unwrap();
     }
 
-    S3Exception::new_err(s)
+    let py_err = S3Exception::new_err(s);
+    let py_err_str = format!("{}", py_err);
+    log_error(&py_err_str);
+    py_err
 }
 
 #[cfg(test)]